pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-manager"
version = "0.1.0"
description = "Modular Telegram automation toolkit"
authors = [{name = "Automation Bot", email = "dev@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyrogram>=2.0.106",
    "pydantic>=1.10",
    "PySide6>=6.5",
    "tgcrypto>=1.2.5",
    "PyYAML>=6.0",
]

[tool.setuptools.packages.find]
include = ["app", "ui"]
